Mcp-swap(fix[lint]): Finish transaction cleanup
why: Transaction hardening left dead helpers and an unchecked close.

what:
- Return sync and close failures from durable directory updates
- Simplify inode formatting and remove unreachable JSONC helpers

+++ mcp/cmd/mcp-swap/physical_identity_unix.go
@@ -19,7 +19,7 @@ func physicalIdentityAt(path string, follow bool) (physicalIdentity, error) {
 		return physicalIdentity{}, err
 	}
 	return physicalIdentity{
-		Device: fmt.Sprint(uint64(info.Dev)),
-		File:   fmt.Sprint(uint64(info.Ino)),
+		Device: fmt.Sprint(info.Dev),
+		File:   fmt.Sprint(info.Ino),
 	}, nil
 }
